Menghapus class user dan menambahkan class employee salary untuk menghindari large class
