+++ src/lib/Employee.java
@@ -5,12 +5,15 @@ import java.time.Month;
 import java.util.LinkedList;
 import java.util.List;
 
-public class Employee {
+public class Employee{
 
 	public Family family;
+	public EmployeeSalary salary;
 
 	private String employeeId;
-	private String idNumber;
+	private String firstName;
+    private String lastName;
+    private String address;
 	
 	private int yearJoined;
 	private int monthJoined;
@@ -20,52 +23,39 @@ public class Employee {
 	private boolean isForeigner;
 	private boolean gender; //true = Laki-laki, false = Perempuan
 	
-	private int monthlySalary;
-	private int otherMonthlyIncome;
-	private int annualDeductible;
 	
-	public Employee(String employeeId, String idNumber, int yearJoined, int monthJoined, int dayJoined, boolean isForeigner, boolean gender) {
+	
+	public Employee(String firstName, String lastName,String employeeId,String address, boolean isForeigner) {
 		this.employeeId = employeeId;
-		this.idNumber = idNumber;
-		this.yearJoined = yearJoined;
-		this.monthJoined = monthJoined;
-		this.dayJoined = dayJoined;
-		this.isForeigner = isForeigner;
-		this.gender = gender;
-		
+		this.firstName = firstName;
+		this.lastName = lastName;
+		this.address = address;
+		this.isForeigner = isForeigner;		
 	}
+
+	public String getEmployeeId(){
+		return employeeId;
+	}
+
+	public boolean getForeigner(){
+		return isForeigner;
+	}
+	public String getFirstName(){
+        return firstName;
+    }
+    public String getLastName(){
+        return lastName;
+    }
+    public String getAddress(){
+        return address;
+    }
 	
 	/**
 	 * Fungsi untuk menentukan gaji bulanan pegawai berdasarkan grade kepegawaiannya (grade 1: 3.000.000 per bulan, grade 2: 5.000.000 per bulan, grade 3: 7.000.000 per bulan)
 	 * Jika pegawai adalah warga negara asing gaji bulanan diperbesar sebanyak 50%
 	 */
 	
-	public void setMonthlySalary(int grade) {	
-		if (grade == 1) {
-			monthlySalary = 3000000;
-			if (isForeigner) {
-				monthlySalary = (int) (3000000 * 1.5);
-			}
-		}else if (grade == 2) {
-			monthlySalary = 5000000;
-			if (isForeigner) {
-				monthlySalary = (int) (3000000 * 1.5);
-			}
-		}else if (grade == 3) {
-			monthlySalary = 7000000;
-			if (isForeigner) {
-				monthlySalary = (int) (3000000 * 1.5);
-			}
-		}
-	}
 	
-	public void setAnnualDeductible(int deductible) {	
-		this.annualDeductible = deductible;
-	}
-	
-	public void setAdditionalIncome(int income) {	
-		this.otherMonthlyIncome = income;
-	}
 	
 	public int getAnnualIncomeTax() {
 		
@@ -78,6 +68,6 @@ public class Employee {
 			monthWorkingInYear = 12;
 		}
 		
-		return TaxFunction.calculateTax(monthlySalary, otherMonthlyIncome, monthWorkingInYear, annualDeductible, family.getSpouseNumber().equals(""), family.getChildIdNumber().size());
+		return TaxFunction.calculateTax(salary.getMonthSalary(), salary.getAddtionalIncome(), monthWorkingInYear, salary.getAnnualDeductible(), family.getSpouseNumber().equals(""), family.getChildIdNumber().size());
 	}
 }
